server/requirements.txt
uvicorn==0.12.2
fastapi==0.61.2
passlib==1.7.4
bcrypt==3.2.0
cryptography==3.2.1
python-jose==3.2.0
python-multipart==0.0.5
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2